+++ frontend/src/App.tsx
@@ -4,6 +4,7 @@ import { JobDetailPage } from "./pages/JobDetailPage";
 import { TestDetailPage } from "./pages/TestDetailPage";
 import { FlakinessPage } from "./pages/FlakinessPage";
 import { ActionRequestPage } from "./pages/ActionRequestPage";
+import { AnalysisTracesPage } from "./pages/AnalysisTracesPage";
 import { Layout } from "./components/Layout";
 import { ManifestProvider } from "./components/ManifestProvider";
 import { CapabilitiesProvider } from "./components/CapabilitiesProvider";
@@ -22,6 +23,7 @@ export default function App() {
               <Route element={<Layout />}>
                 <Route index element={<DashboardPage />} />
                 <Route path="flaky" element={<FlakinessPage />} />
+                <Route path="analysis-traces" element={<AnalysisTracesPage />} />
                 <Route path="job/:jobName" element={<JobDetailPage />} />
                 <Route
                   path="job/:jobName/test/:testName"